+++ src/main/java/com/hobbyhop/domain/club/repository/ClubRepository.java
@@ -1,7 +1,7 @@
 package com.hobbyhop.domain.club.repository;
 
 import com.hobbyhop.domain.club.entity.Club;
-import com.hobbyhop.domain.club.repository.search.ClubRepositoryCustom;
+import com.hobbyhop.domain.club.repository.custom.ClubRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
